asm/c128: Changed int parameters to uintptr.

+++ internal/asm/c128/stubs_amd64.go
@@ -42,10 +42,12 @@ func AxpyIncTo(dst []complex128, incDst, idst uintptr, alpha complex128, x, y []
 func DscalUnitary(alpha float64, x []complex128)
 
 // DscalInc is
-//  for i := 0; i < n; i++ {
-//  	x[i*inc] = complex(real(x[i*inc])*alpha, imag(x[i*inc]))
+//  var ix uintptr
+//  for i := 0; i < int(n); i++ {
+//  	x[ix] = complex(real(x[ix])*alpha, imag(x[ix])*alpha)
+//  	ix += inc
 //  }
-func DscalInc(alpha float64, x []complex128, n, inc int)
+func DscalInc(alpha float64, x []complex128, n, inc uintptr)
 
 // ScalInc is
 //  var ix uintptr
@@ -53,7 +55,7 @@ func DscalInc(alpha float64, x []complex128, n, inc int)
 //  	x[ix] *= alpha
 //  	ix += incX
 //  }
-func ScalInc(alpha complex128, x []complex128, n, inc int)
+func ScalInc(alpha complex128, x []complex128, n, inc uintptr)
 
 // ScalUnitary is
 //  for i := range x {
